+++ domain/src/main/java/woowacourse/movie/domain/movie/PlayingDateTimes.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.movie
 
 import java.time.DayOfWeek
 import java.time.LocalDate
@@ -8,10 +8,6 @@ class PlayingDateTimes(startDate: LocalDate, endDate: LocalDate) { // 뷰를 위
     private val _times: Map<LocalDate, List<LocalTime>> = makeDates(startDate, endDate).associateWith { date -> makeTimes(date) }
     val times: Map<LocalDate, List<LocalTime>>
         get() = _times.toMap()
-    // val startDate: LocalDate
-    //     get() = times.keys.max()
-    // val endDate: LocalDate
-    //     get() = times.keys.min()
 
     private fun makeDates(startDate: LocalDate, endDate: LocalDate): List<LocalDate> {
         var date = startDate
